Dockerization/requirements.txt
pandas
scipy
mysql-connector-python